+++ src/fblas.rs
@@ -10,7 +10,7 @@
 #[no_mangle]
 pub unsafe extern "C" fn CINTdset0(mut n: i32, mut x: *mut f64) {
     let mut i: i32 = 0;
-    i = 0 as i32;
+    i = 0_i32;
     while i < n {
         *x.offset(i as isize) = 0 as f64;
         i += 1;
@@ -26,7 +26,7 @@ pub unsafe extern "C" fn CINTdaxpy2v(
     mut v: *mut f64,
 ) {
     let mut i: i32 = 0;
-    i = 0 as i32;
+    i = 0_i32;
     while i < n {
         *v.offset(i as isize) = a * *x.offset(i as isize) + *y.offset(i as isize);
         i += 1;
@@ -43,26 +43,26 @@ pub unsafe extern "C" fn CINTdmat_transpose(
     let mut i: i32 = 0;
     let mut j: i32 = 0;
     let _k: i32 = 0;
-    j = 0 as i32;
-    while j < n - 3 as i32 {
-        i = 0 as i32;
+    j = 0_i32;
+    while j < n - 3_i32 {
+        i = 0_i32;
         while i < m {
-            *a_t.offset(((j + 0 as i32) * m + i) as isize) =
-                *a.offset((i * n + j + 0 as i32) as isize);
-            *a_t.offset(((j + 1 as i32) * m + i) as isize) =
-                *a.offset((i * n + j + 1 as i32) as isize);
-            *a_t.offset(((j + 2 as i32) * m + i) as isize) =
-                *a.offset((i * n + j + 2 as i32) as isize);
-            *a_t.offset(((j + 3 as i32) * m + i) as isize) =
-                *a.offset((i * n + j + 3 as i32) as isize);
+            *a_t.offset(((j + 0_i32) * m + i) as isize) =
+                *a.offset((i * n + j + 0_i32) as isize);
+            *a_t.offset(((j + 1_i32) * m + i) as isize) =
+                *a.offset((i * n + j + 1_i32) as isize);
+            *a_t.offset(((j + 2_i32) * m + i) as isize) =
+                *a.offset((i * n + j + 2_i32) as isize);
+            *a_t.offset(((j + 3_i32) * m + i) as isize) =
+                *a.offset((i * n + j + 3_i32) as isize);
             i += 1;
             i;
         }
-        j += 4 as i32;
+        j += 4_i32;
     }
     match n - j {
         1 => {
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
                 *a_t.offset((j * m + i) as isize) = *a.offset((i * n + j) as isize);
                 i += 1;
@@ -70,25 +70,25 @@ pub unsafe extern "C" fn CINTdmat_transpose(
             }
         }
         2 => {
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
-                *a_t.offset(((j + 0 as i32) * m + i) as isize) =
-                    *a.offset((i * n + j + 0 as i32) as isize);
-                *a_t.offset(((j + 1 as i32) * m + i) as isize) =
-                    *a.offset((i * n + j + 1 as i32) as isize);
+                *a_t.offset(((j + 0_i32) * m + i) as isize) =
+                    *a.offset((i * n + j + 0_i32) as isize);
+                *a_t.offset(((j + 1_i32) * m + i) as isize) =
+                    *a.offset((i * n + j + 1_i32) as isize);
                 i += 1;
                 i;
             }
         }
         3 => {
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
-                *a_t.offset(((j + 0 as i32) * m + i) as isize) =
-                    *a.offset((i * n + j + 0 as i32) as isize);
-                *a_t.offset(((j + 1 as i32) * m + i) as isize) =
-                    *a.offset((i * n + j + 1 as i32) as isize);
-                *a_t.offset(((j + 2 as i32) * m + i) as isize) =
-                    *a.offset((i * n + j + 2 as i32) as isize);
+                *a_t.offset(((j + 0_i32) * m + i) as isize) =
+                    *a.offset((i * n + j + 0_i32) as isize);
+                *a_t.offset(((j + 1_i32) * m + i) as isize) =
+                    *a.offset((i * n + j + 1_i32) as isize);
+                *a_t.offset(((j + 2_i32) * m + i) as isize) =
+                    *a.offset((i * n + j + 2_i32) as isize);
                 i += 1;
                 i;
             }
@@ -106,26 +106,26 @@ pub unsafe extern "C" fn CINTdplus_transpose(
     let mut i: i32 = 0;
     let mut j: i32 = 0;
     let _k: i32 = 0;
-    j = 0 as i32;
-    while j < n - 3 as i32 {
-        i = 0 as i32;
+    j = 0_i32;
+    while j < n - 3_i32 {
+        i = 0_i32;
         while i < m {
-            *a_t.offset(((j + 0 as i32) * m + i) as isize) +=
-                *a.offset((i * n + j + 0 as i32) as isize);
-            *a_t.offset(((j + 1 as i32) * m + i) as isize) +=
-                *a.offset((i * n + j + 1 as i32) as isize);
-            *a_t.offset(((j + 2 as i32) * m + i) as isize) +=
-                *a.offset((i * n + j + 2 as i32) as isize);
-            *a_t.offset(((j + 3 as i32) * m + i) as isize) +=
-                *a.offset((i * n + j + 3 as i32) as isize);
+            *a_t.offset(((j + 0_i32) * m + i) as isize) +=
+                *a.offset((i * n + j + 0_i32) as isize);
+            *a_t.offset(((j + 1_i32) * m + i) as isize) +=
+                *a.offset((i * n + j + 1_i32) as isize);
+            *a_t.offset(((j + 2_i32) * m + i) as isize) +=
+                *a.offset((i * n + j + 2_i32) as isize);
+            *a_t.offset(((j + 3_i32) * m + i) as isize) +=
+                *a.offset((i * n + j + 3_i32) as isize);
             i += 1;
             i;
         }
-        j += 4 as i32;
+        j += 4_i32;
     }
     match n - j {
         1 => {
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
                 *a_t.offset((j * m + i) as isize) += *a.offset((i * n + j) as isize);
                 i += 1;
@@ -133,25 +133,25 @@ pub unsafe extern "C" fn CINTdplus_transpose(
             }
         }
         2 => {
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
-                *a_t.offset(((j + 0 as i32) * m + i) as isize) +=
-                    *a.offset((i * n + j + 0 as i32) as isize);
-                *a_t.offset(((j + 1 as i32) * m + i) as isize) +=
-                    *a.offset((i * n + j + 1 as i32) as isize);
+                *a_t.offset(((j + 0_i32) * m + i) as isize) +=
+                    *a.offset((i * n + j + 0_i32) as isize);
+                *a_t.offset(((j + 1_i32) * m + i) as isize) +=
+                    *a.offset((i * n + j + 1_i32) as isize);
                 i += 1;
                 i;
             }
         }
         3 => {
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
-                *a_t.offset(((j + 0 as i32) * m + i) as isize) +=
-                    *a.offset((i * n + j + 0 as i32) as isize);
-                *a_t.offset(((j + 1 as i32) * m + i) as isize) +=
-                    *a.offset((i * n + j + 1 as i32) as isize);
-                *a_t.offset(((j + 2 as i32) * m + i) as isize) +=
-                    *a.offset((i * n + j + 2 as i32) as isize);
+                *a_t.offset(((j + 0_i32) * m + i) as isize) +=
+                    *a.offset((i * n + j + 0_i32) as isize);
+                *a_t.offset(((j + 1_i32) * m + i) as isize) +=
+                    *a.offset((i * n + j + 1_i32) as isize);
+                *a_t.offset(((j + 2_i32) * m + i) as isize) +=
+                    *a.offset((i * n + j + 2_i32) as isize);
                 i += 1;
                 i;
             }
@@ -173,18 +173,18 @@ pub unsafe extern "C" fn CINTdgemm_NN1(
     let mut j: i32 = 0;
     let mut kp: i32 = 0;
     let mut bi: f64 = 0.;
-    j = 0 as i32;
+    j = 0_i32;
     while j < n {
-        i = 0 as i32;
+        i = 0_i32;
         while i < m {
             *c.offset((i + ldc * j) as isize) = 0 as f64;
             i += 1;
             i;
         }
-        kp = 0 as i32;
+        kp = 0_i32;
         while kp < k {
             bi = *b.offset((kp + k * j) as isize);
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
                 *c.offset((i + ldc * j) as isize) += *a.offset((i + m * kp) as isize) * bi;
                 i += 1;
@@ -221,12 +221,12 @@ pub unsafe extern "C" fn CINTdgemm_TN(
     let mut j: i32 = 0;
     let mut kp: i32 = 0;
     let mut ci: f64 = 0.;
-    j = 0 as i32;
+    j = 0_i32;
     while j < n {
-        i = 0 as i32;
+        i = 0_i32;
         while i < m {
             ci = 0 as f64;
-            kp = 0 as i32;
+            kp = 0_i32;
             while kp < k {
                 ci += *a.offset((kp + k * i) as isize) * *b.offset((kp + k * j) as isize);
                 kp += 1;
@@ -253,18 +253,18 @@ pub unsafe extern "C" fn CINTdgemm_NT(
     let mut j: i32 = 0;
     let mut kp: i32 = 0;
     let mut bi: f64 = 0.;
-    j = 0 as i32;
+    j = 0_i32;
     while j < n {
-        i = 0 as i32;
+        i = 0_i32;
         while i < m {
             *c.offset((i + m * j) as isize) = 0 as f64;
             i += 1;
             i;
         }
-        kp = 0 as i32;
+        kp = 0_i32;
         while kp < k {
             bi = *b.offset((j + n * kp) as isize);
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
                 *c.offset((i + m * j) as isize) += *a.offset((i + m * kp) as isize) * bi;
                 i += 1;
