+++ src/fmt.rs
@@ -77,9 +77,9 @@ unsafe extern "C" fn fmt1_gamma_inc_like(mut f: *mut f64, mut t: f64, mut m: i32
     }
     *f.offset(m as isize) = s / b;
     i = m;
-    while i > 0 as i32 {
+    while i > 0_i32 {
         b -= 1.0f64;
-        *f.offset((i - 1 as i32) as isize) = (e + t * *f.offset(i as isize)) / b;
+        *f.offset((i - 1_i32) as isize) = (e + t * *f.offset(i as isize)) / b;
         i -= 1;
         i;
     }
@@ -91,15 +91,15 @@ pub unsafe extern "C" fn gamma_inc_like(mut f: *mut f64, mut t: f64, mut m: i32)
     } else {
         let mut i: i32 = 0;
         let mut tt: f64 = sqrt(t);
-        *f.offset(0 as isize) =
+        *f.offset(0_isize) =
             0.8862269254527580136490837416705725913987747280611935641069038949264f64 / tt * erf(tt);
-        if m > 0 as i32 {
+        if m > 0_i32 {
             let mut e: f64 = exp(-t);
             let mut b: f64 = 0.5f64 / t;
-            i = 1 as i32;
+            i = 1_i32;
             while i <= m {
                 *f.offset(i as isize) =
-                    b * ((2 as i32 * i - 1 as i32) as f64 * *f.offset((i - 1 as i32) as isize) - e);
+                    b * ((2_i32 * i - 1_i32) as f64 * *f.offset((i - 1_i32) as isize) - e);
                 i += 1;
                 i;
             }
@@ -122,9 +122,9 @@ unsafe extern "C" fn fmt1_lgamma_inc_like(mut f: *mut f64, mut t: f64, mut m: i3
     }
     *f.offset(m as isize) = s / b;
     i = m;
-    while i > 0 as i32 {
+    while i > 0_i32 {
         b -= 1.0f64;
-        *f.offset((i - 1 as i32) as isize) = (e + t * *f.offset(i as isize)) / b;
+        *f.offset((i - 1_i32) as isize) = (e + t * *f.offset(i as isize)) / b;
         i -= 1;
         i;
     }
@@ -136,15 +136,15 @@ pub unsafe extern "C" fn lgamma_inc_like(mut f: *mut f64, mut t: f64, mut m: i32
     } else {
         let mut i: i32 = 0;
         let mut tt: f64 = sqrtl(t);
-        *f.offset(0 as isize) =
+        *f.offset(0_isize) =
             0.8862269254527580136490837416705725913987747280611935641069038949264 / tt * erfl(tt);
-        if m > 0 as i32 {
+        if m > 0_i32 {
             let mut e: f64 = expl(-t);
             let mut b: f64 = 0.5f64 / t;
-            i = 1 as i32;
+            i = 1_i32;
             while i <= m {
                 *f.offset(i as isize) =
-                    b * ((2 as i32 * i - 1 as i32) as f64 * *f.offset((i - 1 as i32) as isize) - e);
+                    b * ((2_i32 * i - 1_i32) as f64 * *f.offset((i - 1_i32) as isize) - e);
                 i += 1;
                 i;
             }
@@ -154,14 +154,14 @@ pub unsafe extern "C" fn lgamma_inc_like(mut f: *mut f64, mut t: f64, mut m: i32
 #[inline]
 unsafe extern "C" fn _pow(mut base: f64, mut exponent: i32) -> f64 {
     let mut i: i32 = 0;
-    let mut result: f64 = 1 as f64;
-    i = 1 as i32;
+    let mut result: f64 = 1_f64;
+    i = 1_i32;
     while i <= exponent {
         if i & exponent != 0 {
             result *= base;
         }
         base *= base;
-        i <<= 1 as i32;
+        i <<= 1_i32;
     }
     result
 }
@@ -169,13 +169,13 @@ unsafe extern "C" fn _pow(mut base: f64, mut exponent: i32) -> f64 {
 unsafe extern "C" fn _powl(mut base: f64, mut exponent: i32) -> f64 {
     let mut i: i32 = 0;
     let mut result: f64 = 1.0f64;
-    i = 1 as i32;
+    i = 1_i32;
     while i <= exponent {
         if i & exponent != 0 {
             result *= base;
         }
         base *= base;
-        i <<= 1 as i32;
+        i <<= 1_i32;
     }
     result
 }
@@ -205,11 +205,11 @@ pub unsafe extern "C" fn fmt1_erfc_like(mut f: *mut f64, mut t: f64, mut lower: 
     let mut val: f64 = s / b;
     *f.offset(m as isize) = val;
     i = m;
-    while i > 0 as i32 {
+    while i > 0_i32 {
         b -= 1.0f64;
         // e1 /= lower2;
         val = (e - e1 + t * val) / b;
-        *f.offset((i - 1 as i32) as isize) = val;
+        *f.offset((i - 1_i32) as isize) = val;
         i -= 1;
         i;
     }
@@ -221,8 +221,8 @@ pub unsafe extern "C" fn fmt_erfc_like(mut f: *mut f64, mut t: f64, mut lower: f
     }
     let mut i: i32 = 0;
     let mut lower2: f64 = lower * lower;
-    if t * lower2 > 200 as f64 {
-        i = 0 as i32;
+    if t * lower2 > 200_f64 {
+        i = 0_i32;
         while i <= m {
             *f.offset(i as isize) = 0 as f64;
             i += 1;
@@ -237,16 +237,16 @@ pub unsafe extern "C" fn fmt_erfc_like(mut f: *mut f64, mut t: f64, mut lower: f
         let mut val: f64 = 0.8862269254527580136490837416705725913987747280611935641069038949264f64
             / tt
             * (erfc(lower * tt) - erfc(tt));
-        *f.offset(0 as isize) = val;
-        if m > 0 as i32 {
+        *f.offset(0_isize) = val;
+        if m > 0_i32 {
             let mut e: f64 = exp(-t);
             let mut e1: f64 = exp(-t * lower2) * lower;
             let mut b: f64 = 0.5f64 / t;
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
-                val = b * ((2 as i32 * i + 1 as i32) as f64 * val - e + e1);
+                val = b * ((2_i32 * i + 1_i32) as f64 * val - e + e1);
                 e1 *= lower2;
-                *f.offset((i + 1 as i32) as isize) = val;
+                *f.offset((i + 1_i32) as isize) = val;
                 i += 1;
                 i;
             }
@@ -261,7 +261,7 @@ pub unsafe extern "C" fn fmt_lerfc_like(mut f: *mut f64, mut t: f64, mut lower: 
     let mut i: i32 = 0;
     let mut lower2: f64 = lower * lower;
     if t * lower2 > 200.0f64 {
-        i = 0 as i32;
+        i = 0_i32;
         while i <= m {
             *f.offset(i as isize) = 0.0f64;
             i += 1;
@@ -276,16 +276,16 @@ pub unsafe extern "C" fn fmt_lerfc_like(mut f: *mut f64, mut t: f64, mut lower: 
         let mut val: f64 = 0.8862269254527580136490837416705725913987747280611935641069038949264
             / tt
             * (erfcl(lower * tt) - erfcl(tt));
-        *f.offset(0 as isize) = val;
-        if m > 0 as i32 {
+        *f.offset(0_isize) = val;
+        if m > 0_i32 {
             let mut e: f64 = expl(-t);
             let mut e1: f64 = expl(-t * lower2) * lower;
             let mut b: f64 = 0.5f64 / t;
-            i = 0 as i32;
+            i = 0_i32;
             while i < m {
-                val = b * ((2 as i32 * i + 1 as i32) as f64 * val - e + e1);
+                val = b * ((2_i32 * i + 1_i32) as f64 * val - e + e1);
                 e1 *= lower2;
-                *f.offset((i + 1 as i32) as isize) = val;
+                *f.offset((i + 1_i32) as isize) = val;
                 i += 1;
                 i;
             }
@@ -318,11 +318,11 @@ pub unsafe extern "C" fn fmt1_lerfc_like(mut f: *mut f64, mut t: f64, mut lower:
     let mut val: f64 = s / b;
     *f.offset(m as isize) = val;
     i = m;
-    while i > 0 as i32 {
+    while i > 0_i32 {
         b -= 1.0f64;
         // e1 /= lower2;
         val = (e - e1 + t * val) / b;
-        *f.offset((i - 1 as i32) as isize) = val;
+        *f.offset((i - 1_i32) as isize) = val;
         i -= 1;
         i;
     }
